Add explicit CounterAccelerator
This moves the test-defined esptests.TimerAccelerator into
esp.examples.CounterAccelerator (and creates the new esp.examples
package). This makes the examples explicit and allows for non-test
generation of Verilog for the example accelerators.
